vehicleAd capacity type updated

+++ app/src/main/java/com/example/etransportapp/data/model/ad/VehicleAdCreateRequest.kt
@@ -7,6 +7,6 @@ data class VehicleAdCreateRequest(
     val city: String,
     val carrierId: String,
     val vehicleType: String,
-    val capacity: Int,
+    val capacity: Double,
     val adDate: String
 )
